requirements.txt
discord.py==1.7.3
python-dotenv==0.19.1
wikipedia-api==0.5.4
transformers==4.10.0
